+++ paddle/cinn/adt/dim_expr_simplifier.h
@@ -15,12 +15,12 @@
 #pragma once
 
 #include "paddle/cinn/adt/dim_expr.h"
-#include "paddle/cinn/common/dim_expr_simplify.h"
+#include "paddle/pir/dialect/shape/utils/dim_expr_simplify.h"
 
 namespace cinn::adt {
 
 inline DimExpr SimplifyDimExpr(const DimExpr& dim_expr) {
-  return cinn::common::SimplifyDimExpr(dim_expr);
+  return symbol::SimplifyDimExpr(dim_expr);
 }
 
 }  // namespace cinn::adt
