+++ paddle/cinn/common/CMakeLists.txt
@@ -51,8 +51,6 @@ if(WITH_CUDA)
 endif()
 if(NOT CINN_ONLY)
   cinn_cc_test(dim_expr_util_test SRCS dim_expr_util_test.cc DEPS cinncore)
-  cinn_cc_test(dim_expr_simplify_test SRCS dim_expr_simplify_test.cc DEPS
-               cinncore)
   cinn_cc_test(dim_expr_converter_test SRCS dim_expr_converter_test.cc DEPS
                cinncore)
   cinn_cc_test(broadcast_tree_test SRCS broadcast_tree_test.cc DEPS cinncore)
