+++ test/cpp/pir/shape_dialect/CMakeLists.txt
@@ -1,4 +1,5 @@
 paddle_test(symbol_dim_expr_test SRCS symbol_dim_expr_test.cc)
+paddle_test(simplify_dim_expr_test SRCS simplify_dim_expr_test.cc)
 
 if(WITH_CINN)
   paddle_test(shape_optimization_test SRCS shape_optimization_test.cc)
